+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.9] - 2025-09-09
+
+### Changed
+- Changed `lin comments` to `lin comment` for consistency (singular form)
+- Changed `lin projects` to `lin project` for consistency (singular form)
+
 ## [0.1.8] - 2025-09-07
 
 ### Added